tidy up, delete obsolete files
