+++ src/getGhciBasePath.ts
@@ -1,6 +1,5 @@
 import * as os from 'os';
 import * as path from 'path';
-// const path = '/Users/kindohm/.ghcup/bin/ghci';
 
 export const getGhciBasePath = () => {
   return path.join(os.homedir(), ".ghcup", "bin")
